+++ api/routes/recipes.js
@@ -12,8 +12,4 @@ router.post('/', checkAuth, imageUpload.single('recipeImage'), RecipesController
 
 router.get('/:recipeId', RecipesController.getSingleRecipe);
 
-router.patch('/:recipeId', checkAuth, imageUpload.single('recipeImage'), RecipesController.patchRecipe);
-
-router.delete('/:recipeId', checkAuth, RecipesController.deleteRecipe);
-
 module.exports = router;
